// src/pages/fan-matches-page/resources/constants.ts
export enum FanMatchStatsCategory {
  Investors = "Investors",
  Fans = "Fans",
  PPS = "Price Per Share",
  MarketCap = "Market Cap",
  DeltaPPS = "% Delta",
  ShareVolumn = "Share Volume",
  USDCVolumn = "USDC Volume",
  Equity = "Equity",
  TodaysReturn = "Today's Return",
  TotalReturn = "Total Return"
}